First public api propsal (#52)
* Includes following
1) Final proposal for 0.1 public API surface
2) Prefeaturization
3) Splitting train data into train and validate when validation data is null
4) Providing end to end samples one each for regression, binaryclassification and multiclass classification

* Incorporating code review feedbacks

+++ src/AutoML/AutoFitter/AutoFitter.cs
@@ -48,7 +48,7 @@ namespace Microsoft.ML.Auto
             do
             {
                 // get next pipeline
-                var iterationsRemaining = _settings.StoppingCriteria.MaxIterations - _history.Count;
+                var iterationsRemaining = (int)_settings.StoppingCriteria.MaxIterations - _history.Count;
                 var pipeline = PipelineSuggester.GetNextInferredPipeline(_history, columns, _task, iterationsRemaining, _optimizingMetricInfo.IsMaximizing);
 
                 // break if no candidates returned, means no valid pipeline available
